requirements.txt
python-telegram-bot==20.7
openai==1.91.0
gspread==6.2.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
APScheduler==3.11.0
python-dotenv==1.1.0 